Added Portuguese (Portugal pt-PT) Translation
Please note that this is Portuguese from Portugal, and not a change to
Portuguese from Brazil that is already translated.
There are huge diferences on both languages.

I didn't added the Português de Portugal because it's the mother
language and generaly it's written like this.

Translated all the files to have the new language.

Thank you for a nice project :)